requirements.txt
Django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
gunicorn